refactor: small adjustments on the current `create post` workflow
